+++ docs/changelog.md
@@ -6,6 +6,7 @@
 * Preserve cost information (lot dates and lot labels/notes) ([issue #5](https://github.com/beancount/beancount2ledger/issues/5))
 * Avoid adding two prices in hledger ([issue #2](https://github.com/beancount/beancount2ledger/issues/2))
 * Avoid trailing whitespace in account open declarations ([issue #6](https://github.com/beancount/beancount2ledger/issues/6))
+* Fix indentation issue in postings ([issue #8](https://github.com/beancount/beancount2ledger/issues/8))
 * Add documentation
 
 ## 1.0 (2020-07-22)